refactor(hydra-cli): refactoring II
affects: @dzlzv/hydra-cli
