--- packages/hydra-cli/src/helpers/formatter.ts
@@ -1,23 +1,0 @@
-import * as Prettier from 'prettier'
-
-const prettierOptions: Prettier.Options = {
-  parser: 'typescript',
-  endOfLine: 'auto',
-}
-
-export function formatWithPrettier(
-  text: string,
-  options: Prettier.Options = prettierOptions
-): string {
-  let formatted = ''
-  try {
-    formatted = Prettier.format(text, options)
-  } catch (error) {
-    console.error(
-      'There were some errors while formatting with Prettier',
-      error
-    )
-    formatted = text
-  }
-  return formatted
-}
