--- packages/hydra-cli/src/helpers/tsTypes.ts
@@ -1,17 +1,0 @@
-export const fieldTypes: { [key: string]: { [key: string]: string } } = {
-  bool: {
-    tsType: 'boolean',
-  },
-  date: {
-    tsType: 'Date',
-  },
-  int: {
-    tsType: 'number',
-  },
-  float: {
-    tsType: 'number',
-  },
-  string: {
-    tsType: 'string',
-  },
-}
